tests/wpt/web-platform-tests/app-history/navigate-event/navigateerror-ordering-transitionWhile-reentrant.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  let start_href = location.href;
  let events = [];
  let first_onnavigate = true;
  let p2;
  appHistory.onnavigatesuccess = () => events.push("navigatesuccess");
  appHistory.onnavigateerror = t.step_func(() => {
    events.push("navigateerror");
    assert_equals(location.hash, "");
  });
  appHistory.onnavigate = t.step_func(e => {
    assert_equals(location.href, start_href);
    events.push("navigate");
    e.signal.onabort = () => events.push("abort");
    e.transitionWhile(new Promise(resolve => t.step_timeout(resolve, 2)));
    if (first_onnavigate) {
      first_onnavigate = false;
      p2 = appHistory.navigate("#2").then(() => events.push("p2_fulfill"));
    }
  });
  const p1 = promise_rejects_dom(t, 'AbortError', appHistory.navigate("#1")).then(() => {
    events.push("p1_reject");
    assert_equals(location.hash, "#2");
  });
  await p1;
  await p2;
  assert_array_equals(events, ["navigate", "abort", "navigateerror", "navigate", "p1_reject", "navigatesuccess", "p2_fulfill"]);
}, "navigateerror ordering when navigate() is called reentrantly and handled by transitionWhile()");
</script>
